Move interfaces and enums to separate files, fix imports respectively
Auth and Singleplayer components now have "models" folder with their own exported interfaces.
