+++ src/app/auth/models/auth-response-data.ts
@@ -1,0 +1,8 @@
+export interface AuthResponseData {
+  idToken: string,
+  email: string,
+  refreshToken: string,
+  expiresIn: string,
+  localId: string,
+  registered?: boolean
+}
