+++ src/app/auth/models/extra-data.ts
@@ -1,0 +1,6 @@
+import { Statistics } from "src/app/game/singleplayer/models/statistics";
+
+export interface ExtraData {
+  name: string,
+  stats: Statistics
+}
