+++ src/app/game/singleplayer/models/current-action.ts
@@ -1,0 +1,5 @@
+export enum CurrentAction {
+  NewTurn,
+  PotAction,
+  PlantAction
+}
