+++ src/app/game/singleplayer/models/game-constants.ts
@@ -1,0 +1,8 @@
+export class GameConstants {
+  public static readonly COMMUNITY_GARDEN_MAX_HERBS = 5;
+  public static readonly NORMAL_HERBS_COUNT = 9;
+  public static readonly SPECIAL_HERBS_COUNT = 3;
+  public static readonly FINISH_ACTION_SP_ORDER_LENGTH = 3;
+  public static readonly COOKIE_POINTS = 5;
+  public static readonly PLANTABLES_COUNT = 3;
+}
