+++ src/app/game/singleplayer/models/game-score.ts
@@ -1,0 +1,9 @@
+import { Ranks } from "./ranks";
+
+export interface GameScore {
+  points: number,
+  rank: Ranks,
+  nextRankPoints: number,
+  messageLeft: string,
+  messageRight: string
+}
