+++ src/app/game/singleplayer/models/place-in.ts
@@ -1,0 +1,5 @@
+export enum PlaceIn {
+  CommunityGarden,
+  PrivateGarden,
+  DiscardPile,
+}
