+++ src/app/game/singleplayer/models/ranks.ts
@@ -1,0 +1,8 @@
+export enum Ranks {
+  Rank6, // worst
+  Rank5,
+  Rank4,
+  Rank3,
+  Rank2,
+  Rank1 // best
+}
