+++ src/app/game/singleplayer/models/statistics.ts
@@ -1,0 +1,6 @@
+export interface Statistics {
+  topScore: number,
+  gamesPlayed: number,
+  herbsLost: number,
+  perfectScores: number
+}
